requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1.1
django-debug-toolbar==3.1.1
django-nested-inline==0.4.2
django==3.1.2
pillow==7.1.2
psycopg2-binary==2.8.6
pycparser==2.20
pyjwt==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
